requirements.txt
selenium==4.15.0
pytest==7.4.0
allure-pytest==2.13.0
python-dotenv==1.0.0
seletools==1.4.0